client

+++ src/main/java/com/example/dao/CategoryDAO.java
@@ -3,5 +3,6 @@ package com.example.dao;
 import com.example.categories.CourseCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryDAO extends JpaRepository<CourseCategory, Long> {
+public interface CategoryDAO extends JpaRepository<CourseCategory, Integer> {
+
 }
